+++ enterprise/api/v1/views/enterprise_group.py
@@ -18,6 +18,7 @@ from enterprise.api.v1 import serializers
 from enterprise.api.v1.views.base_views import EnterpriseReadWriteModelViewSet
 from enterprise.logging import getEnterpriseLogger
 from enterprise.tasks import send_group_membership_invitation_notification, send_group_membership_removal_notification
+from enterprise.utils import localized_utcnow
 
 LOGGER = getEnterpriseLogger(__name__)
 
@@ -200,6 +201,7 @@ class EnterpriseGroupViewSet(EnterpriseReadWriteModelViewSet):
                 # Extend the list of memberships that need to be created associated with existing Users
                 ent_customer_users = [
                     models.EnterpriseGroupMembership(
+                        activated_at=localized_utcnow(),
                         enterprise_customer_user=ecu,
                         group=group
                     )