+++ integrated_channels/integrated_channel/admin/__init__.py
@@ -99,15 +99,9 @@ class IntegratedChannelAPIRequestLogAdmin(admin.ModelAdmin):
 
     list_display = [
         "endpoint",
-        "enterprise_customer",
+        "enterprise_customer_id",
         "time_taken",
         "status_code",
-    ]
-    list_filter = [
-        "status_code",
-        "enterprise_customer",
-        "endpoint",
-        "time_taken",
     ]
     search_fields = [
         "status_code",
@@ -129,6 +123,10 @@ class IntegratedChannelAPIRequestLogAdmin(admin.ModelAdmin):
     ]
 
     list_per_page = 20
+
+    def get_queryset(self, request):
+        queryset = super().get_queryset(request)
+        return queryset.select_related('enterprise_customer')
 
     class Meta:
         model = IntegratedChannelAPIRequestLogs